test: isolate drift and next round timer logic
